% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\docType{data}
\name{comment1}
\alias{comment1}
\title{Expressions regex}
\format{An object of class \code{character} of length 1.}
\usage{
comment1
}
\description{
différentes expression regex correspondants aux blocs principaux de code SAS
}
\references{
http://www.lsta.upmc.fr/Lejeune/SAS_Lejeune.pdf
}
\keyword{datasets}
