% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_equal_to}
\alias{data_equal_to}
\title{data_equal_to}
\usage{
data_equal_to(code_sas)
}
\description{
detecte la valeur da data et la renvoie, contenur dans une
fonction file.path si elle est associée à une librairie SAS
}
