% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupe.R
\name{decoupe_requete}
\alias{decoupe_requete}
\title{decoupe_requete}
\usage{
decoupe_requete(requete, key_words)
}
\arguments{
\item{requete}{une seule requete sql}

\item{key_words}{: mots clés de découpe (select, from, etc)}
}
\value{
vecteur nommé des blocs, le nom associé correspond aux mots clés
}
\description{
lit une requete sql et renvoie une data.frame avec les mots clés (kw)
et les valeurs associées (sentence)
}
