% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupe.R
\name{decouper_SAS}
\alias{decouper_SAS}
\title{Decoupe code SAS}
\usage{
decouper_SAS(code_sas)
}
\arguments{
\item{code_sas}{: code SAS en entrée, non découpé par lignes}
}
\value{
liste d'élements : place, texte et id
place : localisation du code extrait
texte : code extrait
id : identifiant du code extrait, premier mot ou groupe de mot du bloc,
identifiant la procédure ou le commentaire, etc.
}
\description{
identifie les blocs de procédures/commentaires/étapes data/etc,
les découpe et indique leur position dans le code d'origine
}
