% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_functions.R
\name{paste_str}
\alias{paste_str}
\title{paste_str}
\usage{
paste_str(contenu)
}
\arguments{
\item{contenu}{chaine de caractere contenu de la fonction}
}
\description{
Colle la fonction str() autour d'un contenu
}
