% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr_contents.R
\name{sasr_contents}
\alias{sasr_contents}
\title{sasr_contents}
\usage{
sasr_contents(code_sas)
}
\value{
code R equivalent de la proc contents avec les arguments
}
\description{
transformation de la proc contents en code R equivalent
}
