% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr_libname.R
\name{sasr_libname}
\alias{sasr_libname}
\title{sasr_libname}
\usage{
sasr_libname(code_sas)
}
\arguments{
\item{code_sas}{code en entrée contenant le libname}
}
\value{
expression R associée
}
\description{
transformation l'un libname en une affectation
d'un chemin dans un objet R : Convertit LIBNAME nom chemin en nom <- chemin
}
