% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr_sql.R
\name{sasr_sql}
\alias{sasr_sql}
\title{sasr_sql}
\usage{
sasr_sql(code_sas)
}
\arguments{
\item{code_sas}{code SAS balisé de proc sql; quit;}
}
\value{
la même requeteen R library dplyr
}
\description{
sasr_sql
}
