% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr_sql.R
\name{sql_to_dplyr}
\alias{sql_to_dplyr}
\title{sql_to_dplyr}
\usage{
sql_to_dplyr(code_sql)
}
\arguments{
\item{code_sql}{: chaine de charactère code SQL}
}
\value{
chaine de charactere
}
\description{
sql_to_dplyr
}
