% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripteur.R
\name{transcripteur}
\alias{transcripteur}
\title{Transcripteur}
\usage{
transcripteur(input, output)
}
\arguments{
\item{input}{fichier SAS}

\item{output}{fichier R}
}
\description{
transcrit du code SAS en R
}
