% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_conditions}
\alias{transform_conditions}
\title{transform condition}
\usage{
transform_conditions(chaine)
}
\arguments{
\item{chaine}{chaine de caractère contenant les conditions}
}
\description{
remplace les opérateurs de construction des prédicats pour
qu'ils soient compatibles avec R
}
