% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_functions}
\alias{transform_functions}
\title{transform functions}
\usage{
transform_functions(chaine)
}
\arguments{
\item{chaine}{chaine de charactères contenant la fonction}
}
\description{
remplace les fonctions par défaut SAS ou SQL par des fonctions R
}
