% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_list}
\alias{transform_list}
\title{Transform listes}
\usage{
transform_list(chaine)
}
\arguments{
\item{chaine}{liste SAS au format {l1 l2 l3}}
}
\description{
remplace une liste SAS par son équivalent vectoriel R
}
