% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_path}
\alias{transform_path}
\title{transform path}
\usage{
transform_path(chaine)
}
\arguments{
\item{chaine}{chemin du fichier}
}
\description{
change le chemin d'un fichier de façon compatible à la lecture dans R
En changeant \ en /
}
