% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_pattern}
\alias{transform_pattern}
\title{transform pattern}
\usage{
transform_pattern(chaine)
}
\arguments{
\item{chaine}{chaine de caractères de la clause}
}
\description{
remplace les motifs d'intentification de caractères des LIKES
dans les clauses WHERE d'une requête SQL
}
